move merge_resources into its own mixin
