+++ helix_fhir_client_sdk/responses/get/fhir_get_error_response.py
