+++ helix_fhir_client_sdk/responses/get/fhir_get_list_by_resource_type_response.py
